Change to modal functionality
Added a new response modal to show after a successful surplus or transfer request submit. Restructured modals to not use hidden input fields becasue they can be edited in the inspector (enableing XSS). Instead these values are pushed to the server as soon as they are available and stored in a Session variable